test(interceptor): standardize test describe names
