+++ packages/zimic-interceptor/src/cli/server/token/__tests__/serverTokenCreate.node.test.ts
@@ -22,7 +22,7 @@ import { usingIgnoredConsole } from '@tests/utils/console';
 
 import { clearInterceptorTokens } from './utils';
 
-describe('CLI (server token create)', () => {
+describe('CLI > Server token create', () => {
   const processArgvSpy = vi.spyOn(process, 'argv', 'get');
   const numberOfColorCharactersInTokenValue = 10;
